requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
python-dotenv
mysql-connector-python
joblib